Extract AbstractOTLPIndexingRestIT from OTLPMetricsIndexingRestIT
Extract shared integration test infrastructure (cluster setup, user
credentials, API key creation, index template readiness check) into
a reusable AbstractOTLPIndexingRestIT base class.

This allows logs and traces integration tests to reuse the same test
cluster setup without code duplication.
